requirements.txt
tweepy==4.14.0
python-telegram-bot==20.7
python-dotenv==1.0.0
schedule==1.2.0
selenium==4.15.2
feedparser==6.0.10
requests==2.31.0
psutil==5.9.6 